backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
boto3==1.29.7
pydantic==2.5.0
aiofiles==23.2.1
